# backend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Graphfolio Backend API
  description: Backend API for Graphfolio - Company network visualization
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/stocks:
    get:
      tags:
      - stocks
      summary: Get Sorted Stocks
      description: 'Get sorted stocks list with optional search


        Query params:

        - sort_by: Sort field (ticker, name, price, change_percent, market_cap)

        - limit: Maximum number of stocks to return (default: 50, max: 200)

        - q: Optional search query to filter by ticker or name (case-insensitive)'
      operationId: get_sorted_stocks_api_stocks_get
      parameters:
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: Sort field
          default: ticker
          title: Sort By
        description: Sort field
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of stocks to return (1-200)
          default: 50
          title: Limit
        description: Maximum number of stocks to return (1-200)
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search query (filters by ticker or name)
          title: Q
        description: Search query (filters by ticker or name)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Get Sorted Stocks Api Stocks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stocks/{ticker}:
    get:
      tags:
      - stocks
      summary: Get Stock By Ticker
      description: "Get stock by ticker\n\nReturns full stock information including\
        \ chart data filtered by timeframe.\n\nQuery Parameters:\n- timeframe: Optional\
        \ timeframe filter. Valid options:\n  - 1H: Last 1 hour (limited data availability\
        \ with daily aggregates)\n  - 1D: Last 24 hours\n  - 1W: Last 7 days\n  -\
        \ 1M: Last 30 days\n  - 3M: Last 90 days\n  - 6M: Last 180 days\n  - 1Y: Last\
        \ 365 days\n  - YTD: Year to date (from January 1st)\n  - ALL: All available\
        \ data (default if not specified)\n- before: Optional Unix timestamp (ms).\
        \ Fetch historical data ending before this time.\n  Used for infinite scroll\
        \ / lazy loading of older data."
      operationId: get_stock_by_ticker_api_stocks__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: timeframe
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(1H|1D|1W|1M|3M|6M|1Y|YTD|ALL)$
          - type: 'null'
          description: 'Timeframe filter: 1H, 1D, 1W, 1M, 3M, 6M, 1Y, YTD, ALL'
          title: Timeframe
        description: 'Timeframe filter: 1H, 1D, 1W, 1M, 3M, 6M, 1Y, YTD, ALL'
      - name: before
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Fetch data before this Unix timestamp (ms) for infinite scroll
          title: Before
        description: Fetch data before this Unix timestamp (ms) for infinite scroll
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stocks/{ticker}/basic:
    get:
      tags:
      - stocks
      summary: Get Stock Basic Info
      description: Get basic stock information only (no chart data)
      operationId: get_stock_basic_info_api_stocks__ticker__basic_get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stocks/{ticker}/history:
    get:
      tags:
      - stocks
      summary: Get Stock History
      description: "Get stock price history for sparklines\n\nReturns lightweight\
        \ price history data extracted from chart data.\n\nQuery Parameters:\n- timeframe:\
        \ Optional timeframe filter. Valid options:\n  - 1H: Last 1 hour\n  - 1D:\
        \ Last 24 hours\n  - 1W: Last 7 days\n  - 1M: Last 30 days\n  - 3M: Last 90\
        \ days\n  - 6M: Last 180 days\n  - 1Y: Last 365 days\n  - YTD: Year to date\n\
        \  - ALL: All available data (default)"
      operationId: get_stock_history_api_stocks__ticker__history_get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: timeframe
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(1H|1D|1W|1M|3M|6M|1Y|YTD|ALL)$
          - type: 'null'
          description: 'Timeframe filter: 1H, 1D, 1W, 1M, 3M, 6M, 1Y, YTD, ALL'
          title: Timeframe
        description: 'Timeframe filter: 1H, 1D, 1W, 1M, 3M, 6M, 1Y, YTD, ALL'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/stocks/translations/{ticker}:
    get:
      tags:
      - translations
      summary: Get Translation
      description: 'Get translation for a stock ticker.

        If not found and auto_create=True, creates a pending entry for admin review.'
      operationId: get_translation_api_stocks_translations__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: market
        in: query
        required: true
        schema:
          type: string
          description: Market code (US, TW, JP)
          title: Market
        description: Market code (US, TW, JP)
      - name: name_en
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: English name hint for auto-creation
          title: Name En
        description: English name hint for auto-creation
      - name: auto_create
        in: query
        required: false
        schema:
          type: boolean
          description: Auto-create pending entry if not found
          default: true
          title: Auto Create
        description: Auto-create pending entry if not found
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationPublicResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/translations/{translation_id}:
    get:
      tags:
      - admin
      summary: Get Translation
      description: Get a single translation by ID.
      operationId: get_translation_api_admin_translations__translation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - admin
      summary: Update Translation
      description: Update an existing translation.
      operationId: update_translation_api_admin_translations__translation_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete Translation
      description: Delete a translation.
      operationId: delete_translation_api_admin_translations__translation_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChartDataPoint:
      properties:
        timestamp:
          type: integer
          title: Timestamp
          description: Unix timestamp in milliseconds
        price:
          type: number
          title: Price
          description: Closing price (for backward compatibility)
        date:
          type: string
          title: Date
          description: ISO date string (YYYY-MM-DD)
        open:
          type: number
          title: Open
          description: Opening price
        high:
          type: number
          title: High
          description: Highest price
        low:
          type: number
          title: Low
          description: Lowest price
        close:
          type: number
          title: Close
          description: Closing price
        volume:
          type: integer
          title: Volume
          description: Trading volume
      type: object
      required:
      - timestamp
      - price
      - date
      - open
      - high
      - low
      - close
      - volume
      title: ChartDataPoint
      description: Single OHLCV data point for stock chart
...
    CompanyDetail:
      properties:
        ticker:
          type: string
          title: Ticker
          description: Stock ticker symbol
        name:
          type: string
          title: Name
          description: Company name
        price:
          type: number
          title: Price
          description: Current stock price
        change:
          type: number
          title: Change
          description: Price change (absolute)
        changePercent:
          type: number
          title: Changepercent
          description: Price change (percentage)
        marketCap:
          type: integer
          title: Marketcap
          description: Market capitalization
        revenue:
          type: integer
          title: Revenue
          description: Annual revenue
        pe:
          type: number
          title: Pe
          description: Price-to-earnings ratio
        dividendYield:
          type: number
          title: Dividendyield
          description: Dividend yield
        about:
          type: string
          title: About
          description: Company description
        stats:
          allOf:
          - $ref: '#/components/schemas/StockStats'
          description: Additional statistics
        chartData:
          items:
            $ref: '#/components/schemas/ChartDataPoint'
          type: array
          title: Chartdata
          description: Historical OHLCV data
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
          description: Icon URL (PNG format)
        logo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Logo Url
          description: Logo URL (SVG format)
        icon_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Image
          description: Base64 encoded icon image (PNG format)
        logo_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Logo Image
          description: Base64 encoded logo image (SVG format)
      type: object
      required:
      - ticker
      - name
      - price
      - change
      - changePercent
      - marketCap
      - revenue
      - pe
      - dividendYield
      - about
      - stats
      - chartData
      title: CompanyDetail
      description: Detailed company/stock information
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StockStats:
      properties:
        volume:
          type: integer
          title: Volume
          description: Trading volume
        beta:
          type: number
          title: Beta
          description: Beta coefficient (volatility relative to market)
        volatility:
          type: number
          title: Volatility
          description: Stock volatility
      type: object
      required:
      - volume
      - beta
      - volatility
      title: StockStats
      description: Additional stock statistics
...
    TranslationPublicResponse:
      properties:
        ticker:
          type: string
          title: Ticker
        market:
          type: string
          title: Market
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
      type: object
      required:
      - ticker
      - market
      title: TranslationPublicResponse
      description: Schema for public translation lookup response.
    TranslationResponse:
      properties:
        ticker:
          type: string
          maxLength: 20
          title: Ticker
          description: Stock ticker symbol
        market:
          type: string
          maxLength: 10
          title: Market
          description: Market code (US, TW, JP)
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
          description: English name
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
          description: Chinese Traditional name
        id:
          type: integer
          title: Id
        translation_status:
          type: string
          title: Translation Status
        last_updated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Updated By
        last_updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Updated At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - ticker
      - market
      - id
      - translation_status
      title: TranslationResponse
      description: Schema for translation response.
    TranslationUpdate:
      properties:
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
        translation_status:
          anyOf:
          - type: string
            enum:
            - pending
            - approved
            - auto
          - type: 'null'
          title: Translation Status
      type: object
      title: TranslationUpdate
      description: Schema for updating an existing translation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer